+++ arviz/tests/base_tests/test_plots_matplotlib.py
@@ -261,7 +261,7 @@ def test_plot_trace_legend(compact, combined):
     axes = plot_trace(
         idata, var_names=["home", "atts_star"], compact=compact, combined=combined, legend=True
     )
-    assert axes[0, 0].get_legend()
+    assert axes[0, 1].get_legend()
     compact_legend = axes[1, 0].get_legend()
     if compact:
         assert axes.shape == (2, 2)